Auth

+++ app/Http/Controllers/BookController.php
@@ -40,6 +40,23 @@ class BookController extends Controller
      */
     public function store(Request $request)
     {
+        $formFields =[
+            'title' =>'required|string|max:80',
+            'author'=>'required|string|max:80',
+            'pages'=>'required|int|min:5|max:1000',
+            'ISBN'=>'required|string|max:13|min:13',
+            'front'=>'required|max:8000|mimes:jpg,png,jpeg,webp',
+        ];
+
+        $error = [
+            'title.required'=>'El Título es requerido, no debe contener mas de 80 caracteres',
+            'author.required'=>'El Autor es requerido, su nombre no debe ser mayor a 80 caracteres',
+            'pages.required'=>'El número de paginas son requeridas, debe estar en un rango de 5 y 1000',
+            'ISBN.required'=>'El ISBN es requerido, debe tener 13 carácteres',
+            'front.required'=>'La portada es requerida, no debe ser superior a 8Mb',
+        ];
+
+        $this->validate($request,$formFields,$error);
         $bookCase =  request()->except('_token');
 
         if($request->hasFile('front')){
@@ -56,8 +73,11 @@ class BookController extends Controller
      * @param  \App\Models\Book  $book
      * @return \Illuminate\Http\Response
      */
-    public function show(Book $id)
+    public function show($id)
     {
+        $bookCase = Book::findOrFail($id);
+
+        return view('books.show', compact('bookCase'));
 
     }
 
@@ -84,9 +104,29 @@ class BookController extends Controller
      */
     public function update(Request $request,$id)
     {
+        $formFields =[
+            'title' =>'required|string|max:80',
+            'author'=>'required|string|max:80',
+            'pages'=>'required|int|min:5|max:10000',
+            'ISBN'=>'required|string|max:13|min:13',
+        ];
+
+        $error = [
+            'title.required'=>'El Título es requerido, no debe contener mas de 80 caracteres',
+            'author.required'=>'El Autor es requerido, su nombre no debe ser mayor a 80 caracteres',
+            'pages.required'=>'El número de paginas son requeridas, debe estar en un rango de 5 y 1000',
+            'ISBN.required'=>'El ISBN es requerido, debe tener 13 carácteres',
+            'ISBN.unique'=>'Este ISBN ya se encuentra registrado, por favor ingrese otro',
+            'front.required'=>'La portada es requerida, no debe ser superior a 8Mb',
+        ];
+
+        $this->validate($request, $formFields,$error);
         $book =request()->except(['_token', '_method']);
 
         if($request->hasFile('front')){
+            $formFields =[ 'front'=>'required|max:8000|mimes:jpg,png,jpeg,webp',];
+            $error = ['required'=> 'La portada es requerida'];
+
             $bookCase = Book::findOrFail($id);
             Storage::delete(['public/', $bookCase->front]);
             $book['front']= $request->file('front')->store('uploads', 'public');
@@ -94,7 +134,8 @@ class BookController extends Controller
 
         Book::where('id', '=', $id)->update($book);
         $bookCase = Book::findOrFail($id);
-       return view('books.edit', compact('bookCase') );
+        return redirect('books')->with('msg', 'El libro se ha editado correctamente');
+
     }
 
     /**
@@ -107,6 +148,9 @@ class BookController extends Controller
     {
         $bookCase = Book::find($id);
 
+        if(!$bookCase->on_stock){
+            return view('books.delete');
+        }
         if(Storage::delete('public/'. $bookCase->front)){
             Book::destroy($id);
         }
